moving files referencing generated under /internal
