+++ sdk/communication/communication-chat/.vscode/launch.json
@@ -15,7 +15,7 @@
         "--timeout",
         "999999",
         "--colors",
-        "${workspaceFolder}/test/*.spec.ts"
+        "${workspaceFolder}/test/**/*.spec.ts"
       ],
       "env": {
         "TS_NODE_COMPILER_OPTIONS": "{\"module\": \"commonjs\"}",
