+++ sdk/communication/communication-chat/test/internal/chatClient.mocked.spec.ts
@@ -6,7 +6,7 @@ import { assert } from "chai";
 import { ChatClient, CreateChatThreadRequest } from "../../src";
 import * as RestModel from "../../src/generated/src/models";
 import { apiVersion } from "../../src/generated/src/models/parameters";
-import { baseUri, generateToken } from "./utils/connectionUtils";
+import { baseUri, generateToken } from "../public/utils/connectionUtils";
 import {
   AzureCommunicationTokenCredential,
   CommunicationUserIdentifier
