+++ sdk/communication/communication-chat/test/internal/chatThreadClient.mocked.spec.ts
@@ -16,7 +16,7 @@ import {
 } from "../../src";
 import * as RestModel from "../../src/generated/src/models";
 import { apiVersion } from "../../src/generated/src/models/parameters";
-import { baseUri, generateToken } from "./utils/connectionUtils";
+import { baseUri, generateToken } from "../public/utils/connectionUtils";
 import {
   generateHttpClient,
   createChatThreadClient,
